+++ fns-hl7-pipeline/fn-receiver-debatcher/src/main/kotlin/gov/cdc/dex/hl7/ReceiverProcessMetadata.kt
@@ -1,10 +1,10 @@
 package gov.cdc.dex.hl7
 
-import gov.cdc.dex.azure.EventHubMetadata
-import gov.cdc.dex.metadata.ProcessMetadata
+import gov.cdc.dex.metadata.EventGridStageMetadata
 
-data class ReceiverProcessMetadata (override val status: String?,  @Transient val eventHubMD: EventHubMetadata): ProcessMetadata(
-    RECEIVER_PROCESS, RECEIVER_VERSION , status, eventHubMD, listOf()) {
+data class ReceiverProcessMetadata (override val status: String?,
+                               override val eventTimestamp: String): EventGridStageMetadata (
+    RECEIVER_PROCESS, RECEIVER_VERSION , status, listOf(), eventTimestamp ) {
     companion object  {
         const val RECEIVER_PROCESS = "RECEIVER"
         const val RECEIVER_VERSION = "1.0.0"