+++ fns-hl7-pipeline/fn-receiver-debatcher/src/test/kotlin/gov/cdc/dex/hl7/FunctionTest.kt
@@ -1,12 +1,13 @@
 package gov.cdc.dex.hl7
 
-import com.google.gson.JsonObject
+import com.microsoft.azure.functions.ExecutionContext
 import com.microsoft.azure.functions.OutputBinding
 import gov.cdc.dex.azure.EventHubMetadata
 import gov.cdc.dex.metadata.DexEventPayload
 import gov.cdc.dex.metadata.HL7MessageType
 import org.junit.jupiter.api.Test
-
+import org.junit.jupiter.api.parallel.Execution
+import java.util.logging.Logger
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
@@ -16,11 +17,9 @@ class FunctionTest {
     private fun processFile(filename: String): DexEventPayload? {
         println("Start processing $filename ")
         val text = this::class.java.getResource("/$filename").readText()
-        val messages = listOf(text)
-        val eventHubMDList = listOf(EventHubMetadata(1, 99, "", ""))
 
         val function = Function()
-        return function.eventHubProcessor(messages, eventHubMDList)
+        return function.processEventGrid(text, getExecutionContext())
     }
 
     @Test
@@ -29,7 +28,7 @@ class FunctionTest {
         //println(dexEvtPayLoad)
         assertNotNull(dexEvtPayLoad)
         assertEquals(HL7MessageType.ELR, dexEvtPayLoad.messageInfo.type)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
+        assertNotNull(dexEvtPayLoad.metadata.stage)
         assertEquals("RECEIVED", dexEvtPayLoad.summary.currentStatus)
         assertNull(dexEvtPayLoad.summary.problem)
     }
@@ -39,31 +38,18 @@ class FunctionTest {
         //println(dexEvtPayLoad)
         assertNotNull(dexEvtPayLoad)
         assertEquals(HL7MessageType.CASE,dexEvtPayLoad.messageInfo.type)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
+        assertNotNull(dexEvtPayLoad.metadata.stage)
         assertEquals("RECEIVED", dexEvtPayLoad.summary.currentStatus)
         assertNull(dexEvtPayLoad.summary.problem)
     }
-//    @Test
-//    fun process_ErrorPath() {
-//        var dexEvtPayLoad = processFile("ERROR_message.txt")
-//        println(dexEvtPayLoad)
-//    }
-    @Test
-    fun process_NoMetadata() {
-        val dexEvtPayLoad = processFile("NoMetadataFile.txt")
-        assertNotNull(dexEvtPayLoad)
-        assertEquals(HL7MessageType.UNKNOWN,dexEvtPayLoad.messageInfo.type)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
-        assertEquals("REJECTED", dexEvtPayLoad.summary.currentStatus)
-        assertNotNull(dexEvtPayLoad.summary.problem)
-    }
+
     @Test
     fun process_BatchMessage() {
         val dexEvtPayLoad = processFile("BatchMessage.txt")
         assertNotNull(dexEvtPayLoad)
         assertEquals(HL7MessageType.CASE,dexEvtPayLoad.messageInfo.type)
         assertEquals("BATCH", dexEvtPayLoad.metadata.provenance.singleOrBatch)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
+        assertNotNull(dexEvtPayLoad.metadata.stage)
         assertEquals("RECEIVED", dexEvtPayLoad.summary.currentStatus)
         assertNull(dexEvtPayLoad.summary.problem)
     }
@@ -72,21 +58,12 @@ class FunctionTest {
         val dexEvtPayLoad = processFile("InvalidMessage.txt")
         assertNotNull(dexEvtPayLoad)
         assertEquals(HL7MessageType.CASE,dexEvtPayLoad.messageInfo.type)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
+        assertNotNull(dexEvtPayLoad.metadata.stage)
         assertEquals("SINGLE", dexEvtPayLoad.metadata.provenance.singleOrBatch)
         assertEquals("REJECTED", dexEvtPayLoad.summary.currentStatus)
         assertNotNull(dexEvtPayLoad.summary.problem)
     }
 
-    @Test
-    fun process_ELRWithBlanks() {
-        val dexEvtPayLoad = processFile("CovidELRWithBlanks.txt")
-        assertNotNull(dexEvtPayLoad)
-        assertEquals(HL7MessageType.ELR, dexEvtPayLoad.messageInfo.type)
-        assertNotNull(dexEvtPayLoad.metadata.processes)
-        assertEquals("RECEIVED", dexEvtPayLoad.summary.currentStatus)
-        assertNull(dexEvtPayLoad.summary.problem)
-    }
 
     private fun <T> getOutputBinding(): OutputBinding<T> {
         return object : OutputBinding<T> {
@@ -97,6 +74,21 @@ class FunctionTest {
 
             override fun setValue(p0:T?) {
                 inner = p0
+            }
+        }
+    }
+    private fun getExecutionContext():ExecutionContext {
+        return object :ExecutionContext {
+            override fun getLogger(): Logger {
+                return Logger.getLogger(Function:: class.java.name)
+            }
+
+            override fun getInvocationId():String {
+                return "null"
+            }
+
+            override fun getFunctionName():String {
+                return "HL7_JSON_LAKE_TRANSFORMER"
             }
         }
     }