+++ local_libs/lib-bumblebee/src/test/kotlin/gov/cdc/hl7/bumblebee/TestHL7Transform.kt
@@ -167,8 +167,8 @@ class TestHL7Transform {
     @Test
     fun testHL7Transformer() {
 //        val message = this::class.java.getResource("/COVID.txt").readText()
-        val message = this::class.java.getResource("/testFile.hl7").readText()
-        val gson = GsonBuilder().serializeNulls().create()
+        val message = this::class.java.getResource("/testMessage.txt").readText()
+        val gson = GsonBuilder().create()
 
         val xformer = HL7JsonTransformer.getTransformerWithResource(message, "PhinGuideProfile.json")
         val fullHL7 = xformer.transformMessage()