+++ local_libs/lib-dex-commons/src/main/kotlin/gov/cdc/dex/azure/DedicatedEventHubSender.kt
@@ -1,5 +1,6 @@
 package gov.cdc.dex.azure
 
+import com.azure.core.amqp.exception.AmqpException
 import com.azure.messaging.eventhubs.EventData
 import com.azure.messaging.eventhubs.EventHubClientBuilder
 import com.azure.messaging.eventhubs.EventHubProducerClient
@@ -13,32 +14,49 @@ class DedicatedEventHubSender( evHubConnStr: String, evHubTopicName: String) {
         producer.close()
     }
 
-    fun send(message: String) {
-        send(listOf(message))
+    fun send(message: String) : List<Int> {
+        return send(listOf(message))
     }
 
-    fun send(messages: List<String>) {
-
+    fun send(messages: List<String>) : List<Int> {
         var eventDataBatch = producer.createBatch()
-        messages.forEach { msg ->
+        // if any error out due to size, note which ones.
+        // any other errors will propagate back to the caller
+        val errors: MutableList<Int> = mutableListOf()
+        messages.forEachIndexed { index,  msg ->
             // try to add the event from the array to the batch
-            if (!eventDataBatch.tryAdd(EventData(msg))) {
-                // if the batch is full, send it and then create a new batch
-                producer.send(eventDataBatch)
-                eventDataBatch = producer.createBatch()
+            val eventData = EventData(msg)
+            if (eventData.body.size > eventDataBatch.maxSizeInBytes) {
+                errors.add(index)
+            } else {
+                if (!eventDataBatch.tryAdd(eventData)) {
+                    // if the batch has data, send it and then create a new batch
+                    if (eventDataBatch.count > 0) {
+                        producer.send(eventDataBatch)
+                        eventDataBatch = producer.createBatch()
 
-                // Try to add that event that couldn't fit before.
-                if (!eventDataBatch.tryAdd(EventData(msg))) {
-                    throw IllegalArgumentException(
-                        "Event is too large for an empty batch. Max size: "
-                                + eventDataBatch.maxSizeInBytes
-                    )
+                        // Try to add that event that couldn't fit before.
+                        try {
+                            if (!eventDataBatch.tryAdd(eventData)) {
+                                // Event is too large for empty batch
+                                errors.add(index)
+                            } // .if
+                        } catch (e: AmqpException) {
+                            // tryAdd will (sometimes?) throw an AmqpException if the
+                            // event data exceeds the max batch size.
+                            errors.add(index)
+                        }
+                    } else {
+                        // Event is too large for empty batch
+                        errors.add(index)
+                    }
                 } // .if
-            } // .if
+            }
         } // .for
         // send the last batch of remaining events
         if (eventDataBatch.count > 0) {
             producer.send(eventDataBatch)
         }
+        return errors
     } // .send
 }